FEAT: Tarea 2 Aplicaciones distribuidas

+++ src/EliminarTask.js
@@ -1,31 +1,55 @@
-const {DynamoDB} = require('aws-sdk');
+const {
+    DynamoDB
+} = require('aws-sdk');
 
 exports.eliminarTask = async (event) => {
     const dynamoDB = new DynamoDB.DocumentClient();
-    const { id } = event.pathParameters; // Obtener el id desde la ruta
+    const {
+        id
+    } = event.pathParameters; // Obtener el id desde la ruta
+    if (!id) {
+        return {
+            statusCode: 400,
+            body: JSON.stringify({
+                message: 'ID de tarea no proporcionado',
+            })
+        };
+    }
+
     try {
         const response = await dynamoDB.delete({
             TableName: 'usersTable',
             Key: {
-                id, // Usar el id obtenido de los parámetros de la ruta
+                id
             },
-            ReturnValues: 'ALL_OLD' // Devuelve el elemento eliminado
+            ReturnValues: 'ALL_OLD'
         }).promise();
+
+        if (!response.Attributes) {
+            return {
+                statusCode: 404,
+                body: JSON.stringify({
+                    message: 'Tarea no encontrada',
+                })
+            };
+        }
 
         return {
             statusCode: 200,
             body: JSON.stringify({
                 message: 'Tarea eliminada con éxito',
-                data: response.Attributes // Devuelve el elemento eliminado
+                data: response.Attributes
             })
         };
     } catch (error) {
+        console.error("Error al eliminar tarea:", error);
         return {
             statusCode: 500,
             body: JSON.stringify({
-                message: 'Error al eliminar tarea',
+                message: 'Error interno al eliminar tarea',
                 error: error.message
             })
         };
     }
+
 };
